# src/main/resources/docs/Swagger-doc.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Gerenciamento de Estoque",
    "description": "API restful criado para facilitar o estoque de produtos.",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Servidor local de desenvolvimento"
    },
    {
      "url": "http://estoque-production.up.railway.app",
      "description": "Servidor de produção"
    }
  ],
...
  "paths": {
...
    "/api/fornecedor/atualizar/{id}": {
      "put": {
        "tags": [
          "Fornecedores"
        ],
        "summary": "Atualizar fornecedor por ID",
        "description": "Atualiza um fornecedor pelo ID fornecido.",
        "operationId": "update_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FornecedorDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Fornecedor atualizado com sucesso.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Violação na integridade dos dados",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Fornecedor não encontrado.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/fornecedor/cadastrar": {
      "post": {
        "tags": [
          "Fornecedores"
        ],
        "summary": "Cadastrar um novo fornecedor",
        "description": "Cadastra um novo fornecedor com base nos dados fornecidos.",
        "operationId": "create_3",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FornecedorDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Fornecedor cadastrado com sucesso!",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Violação na integridade dos dados",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/produtoSaida/{id}": {
      "get": {
        "tags": [
          "Produto Saida"
        ],
        "summary": "Buscar saída de produto por ID",
        "description": "Recupera uma saída de produto pelo seu ID.",
        "operationId": "findById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Saída de produto encontrada com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoSaidaGetDTO"
                }
              }
            }
          },
          "404": {
            "description": "Saída de produto não encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoSaidaGetDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/produtoPerda/{id}": {
      "get": {
        "tags": [
          "Produto Perda"
        ],
        "summary": "Buscar perda de produto por ID",
        "description": "Recupera uma perda de produto pelo seu ID.",
        "operationId": "findById_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Perda de produto encontrada com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoPerdaGetDTO"
                }
              }
            }
          },
          "404": {
            "description": "Perda de produto não encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoPerdaGetDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/produtoEntrada/{id}": {
      "get": {
        "tags": [
          "Produto Entrada"
        ],
        "summary": "Buscar entrada de produto por ID",
        "description": "Recupera uma entrada de produto pelo seu ID.",
        "operationId": "findById_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Entrada de produto encontrada com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoEntradaGetDTO"
                }
              }
            }
          },
          "404": {
            "description": "Entrada de produto não encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoEntradaGetDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/produtoCapa/{id}": {
      "get": {
        "tags": [
          "ProdutoCapa"
        ],
        "summary": "Buscar produto por ID",
        "description": "Recupera um produto pelo seu ID.",
        "operationId": "findById_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produto encontrado com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoCapaGetDTO"
                }
              }
            }
          },
          "404": {
            "description": "Produto não encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoCapaGetDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/fornecedor/{id}": {
      "get": {
        "tags": [
          "Fornecedores"
        ],
        "summary": "Buscar fornecedor por ID",
        "description": "Recupera um fornecedor pelo ID fornecido.",
        "operationId": "findById_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fornecedor encontrado com sucesso.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Fornecedor não encontrado.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/fornecedor/deletar/{id}": {
      "delete": {
        "tags": [
          "Fornecedores"
        ],
        "summary": "Deletar fornecedor por ID",
        "description": "Exclui um fornecedor pelo ID fornecido.",
        "operationId": "delete_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fornecedor excluído com sucesso!",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Fornecedor não encontrado.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "FornecedorDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "empresa": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "tipoEmpresa": {
            "type": "string",
            "enum": [
              "FORNECEDOR",
              "CLIENTE"
            ]
          },
          "email": {
            "type": "string"
          },
          "telefone": {
            "type": "string"
          },
          "endereco": {
            "type": "string"
          },
          "ativo": {
            "type": "boolean"
          }
        }
      },
      "ProdutoSaidaGetDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "dataSaida": {
            "type": "string",
            "format": "date-time"
          },
          "quantidade": {
            "type": "integer",
            "format": "int64"
          },
          "retiradoPor": {
            "type": "string"
          },
          "setor": {
            "type": "string",
            "enum": [
              "MARKETING",
              "ESCRITORIO",
              "USO_INTERNO",
              "SERVICOS_GERAIS"
            ]
          },
          "observacao": {
            "type": "string"
          }
        }
      },
      "ProdutoPerdaGetDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "data": {
            "type": "string",
            "format": "date-time"
          },
          "quantidade": {
            "type": "integer",
            "format": "int64"
          },
          "motivo": {
            "type": "string"
          }
        }
      },
      "ProdutoEntradaGetDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "numeroNota": {
            "type": "integer",
            "format": "int64"
          },
          "dataPedido": {
            "type": "string",
            "format": "date-time"
          },
          "dataEntrega": {
            "type": "string",
            "format": "date-time"
          },
          "precoCompra": {
            "type": "number",
            "format": "double"
          },
          "quantidade": {
            "type": "integer",
            "format": "int64"
          },
          "observacao": {
            "type": "string"
          }
        }
      },
      "FornecedorGetDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "empresa": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          }
        }
      },
      "ProdutoCapaGetDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "desc": {
            "type": "string"
          },
          "tipoProduto": {
            "type": "string",
            "enum": [
              "MATERIA_PRIMA",
              "PRODUTO_PRONTO",
              "DIVERSOS"
            ]
          },
          "medidaUnidade": {
            "type": "string",
            "enum": [
              "UNIDADE",
              "PACOTE",
              "KILO",
              "PECA",
              "METRO",
              "GRAMAS",
              "CAIXA",
              "SACO",
              "PAR",
              "MILHEIRO",
              "MILIGRAMAS"
            ]
          },
          "fornecedor": {
            "$ref": "#/components/schemas/FornecedorGetDTO"
          },
          "entradas": {
            "type": "number",
            "format": "double"
          },
          "saidas": {
            "type": "number",
            "format": "double"
          },
          "perdas": {
            "type": "number",
            "format": "double"
          },
          "saldo": {
            "type": "number",
            "format": "double"
          },
          "valorCompra": {
            "type": "number",
            "format": "double"
          },
          "valorTotal": {
            "type": "number",
            "format": "double"
          },
          "minimo": {
            "type": "integer",
            "format": "int64"
          },
          "maximo": {
            "type": "integer",
            "format": "int64"
          },
          "resuprimento": {
            "type": "string",
            "enum": [
              "ESTOQUE_NEGATIVO",
              "SALDO_ZERADO",
              "COMPRAR_AGORA",
              "QUANTIDADE_IDEAL",
              "PRODUTO_EXCESSO"
            ]
          },
          "ativo": {
            "type": "boolean"
          }
        }
      }
    }
  },
...
}